Changes migrations to use big ints

+++ database/migrations/add_stripe_billing_columns_to_owner_table.php
@@ -15,7 +15,7 @@ class AddStripeBillingColumnsToOwnerTable extends Migration
     {
         Schema::table(config('stripe-billing.tables.owner'), function (Blueprint $table) {
             $table->string('stripe_id')->nullable()->collation('utf8mb4_bin');
-            $table->unsignedInteger('default_card_id')->nullable();
+            $table->unsignedBigInteger('default_card_id')->nullable();
         });
     }
