+++ database/migrations/create_cards_table.php
@@ -15,7 +15,7 @@ class CreateCardsTable extends Migration
     {
         Schema::create(config('stripe-billing.tables.cards'), function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('owner_id');
+            $table->unsignedBigInteger('owner_id');
             $table->string('stripe_card_id');
             $table->string('brand', 30);
             $table->string('last_4', 8);
